Fix selectedIndex UI/UX effect on WatchList
