+++ src/components/protectedRoutes/Cryptos.js
@@ -5,24 +5,23 @@ import { ListItem, ListItemIcon, ListItemText, Avatar } from "@material-ui/core"
 //
 const Cryptos = (props) => {
 
-  const handleListItemClick = (event, index) => {
-    console.log("event id", event.currentTarget.id);
-    props.handleSelectedIndex(index);
-    props.cryptoHandler(event.currentTarget.id);
-  };
+  const clickHandler = (eventTargetId, index) => {
+    props.selectedHandler(index);
+    props.cryptoHandler(eventTargetId);
+  }
 
   return (
     <ListItem
       button
       id={props.id}
       key={props.id}
-      selected={props.selectedIndex === props.id}
-      onClick={(event) => handleListItemClick(event, props.i)}
+      selected={props.selectedIndex === props.index}
+      onClick={event => clickHandler(event.currentTarget.id, props.index)}
     >
       <ListItemIcon>
-        <Avatar alt="bitcoin-logo" src={`/images/${props.crypto.name}-logo.png`} />
+        <Avatar alt="bitcoin-logo" src={`/images/${props.id}-logo.png`} />
       </ListItemIcon>
-      <ListItemText primary={props.crypto.name} />
+      <ListItemText primary={props.id} />
     </ListItem>
   )
 }
